// HealthTracker_FE/src/types/PatientData.tsx
export interface VitalsData {
    date: string;
    heart_rate: number;
    pulse_rate: number;
    systolic_bp: number;
    diastolic_bp: number;
    oxygen_saturation: number;
    respiration_rate: number;
    body_temperature: number;
    cholesterol_level: number;
    weight: number;
    sleep_duration: number;
    calories_burned: number;
    steps_count: number;
    physician_note: string;
  }